oracle/spec: Set minimum kernel version for UEK5
We only support the packages derived from the RDMA Core project, on
OL7+UEK5 and later.

These packages should be compatible with UEK4 but have not been
thoroughly tested so we attempt to block their use on systems without
a UEK5 kernel.

Note: The check performed by RPM only blocks installation if a UEK5
      kernel >= 4.14.35-1844 is not installed on the system. There is no
      way to ensure that correct UEK5 kernel is actually RUNNING.

Orabug: 29391672

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>