oracle: Add buildrpm and link spec file
THIS SHOULD NOT BE SUBMITTED UPSTREAM. DO NOT SQUASH!

'buildrpm' is required by the Oracle auto build system. Rather than
deviate from the upstream convention of having a per-distro directory
for distro-specific files, we add buildrpm and symlink the RPM spec
file so the build system can find it witin the 'oracle' directory.

buildrpm/
└── rdma-core.spec -> ../oracle/rdma-core.spec

An enhancement request has been opened to allow specifying the location
of the spec file rather than defaulting to 'buildrpm'. This provides
better alignment with the methods used by upstream projects. If/when
that ER is implemented, this commit will be reverted.

The packages derived from this project will only be built for OL7 and
will be restricted to OL7 with the UEK5 kernel. They may be compatible
with OL7+UEK4 but that has not been tested.

THIS SHOULD NOT BE SUBMITTED UPSTREAM. DO NOT SQUASH!

Orabug: 26829960

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>