verbs: Introduce ibv_advise_mr verb
Introduce a new verb named ibv_advise_mr(), it includes:
- The application interface.
- The command interface with the kernel.

The purpose of this patch is to give an advice to the kernel
about an address range belongs to a memory region.

Performance - applications that are aware of future access patterns can
use this verb in order to leverage this knowledge to improve system or
application performance.

One example is for pre-fetching a range part of an on-demand paging MR.

On-demand-paging (ODP) is a technique to alleviate much of the
shortcomings of memory registration.  The main idea is that applications
no longer pin down the underlying physical pages of the address space
and need not track the validity of the mappings.  Rather, the HCA
requests the latest translations from the OS when pages are not present,
and the OS invalidates translations which are no longer valid due to
mapping changes.  Pre-fetching ODP address ranges, therefore make them
present before the actual IO is conducted.  This would provide a way to
reduce latency by overlapping paging-in and either compute time or IO to
other ranges.

Signed-off-by: Moni Shoua <monis@mellanox.com>
Signed-off-by: Aviad Yehezkel <aviadye@mellanox.com>
Reviewed-by: Guy Levi <guyle@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>

Orabug: 29180408

(cherry picked from commit b341e606265f7f2f2791456c90d8589658f0fddc)
cherry-pick-repo=linux-rdma/rdma-core.git

Conflicts:
	libibverbs/driver.h
	libibverbs/libibverbs.map.in
	libibverbs/verbs.h
	libibverbs/verbs_api.h

The conflicts were resolved by removing the changes that were introduced
post v19.1.

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Signed-off-by: Aron Silverton <aron.silverton@oracle.com>

Orabug: 29410510

Rebase from RDMA Core 19.2 -> 20.2.

(cherry picked from commit f3b86de9)
cherry-pick-repo=linux-git/RDMA/rdma-core.git

Conflicts (19.2 -> 20.2):
	libibverbs/driver.h
	libibverbs/libibverbs.map.in
	libibverbs/verbs.h
	libibverbs/verbs_api.h

All the conflicts were a result of changes that occurred between 19.2
and 20 upstream. We just resolved them by making sure that the changes
from 20 were applied correctly.

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>