oracle: Load RDMAIP kernel module using hot-plug
Load the 'resilient_rdmaip' kernel module using the upstream kernel-boot
hot-plug framework.

NB: In the past, changes were made to the network service
(/etc/rc.d/init.d/network) to make it aware of the RDMAIP module. Those
changes include:

1. Disabling active/active bonding when the network service is stopped
   and enabling active/active bonding when the network service is
   started. In this context, stopping the network service means running
   'ifdown' on all interfaces so it is not clear why RDMAIP would still
   be atempting to do anything with the interfaces when they are down.
   RDMAIP should use netlink or other facilities to be aware of the
   interface state.

2. Triggering RDMAIP to run its bonding code when interfaces are up.
   This is apparently done to avoid some sort of delay in how long it
   might otherwise take RDMAIP to bond the interfaces. Again, this
   should be done automatically by having RDMAIP use netlink to be aware
   of interfaces state changes.

If/when networking configuration is done via systemd-networkd, this
functionality will need to be reimplemented using systemd if it is still
required. It would be better to fix 'resilient_rdmaip' so that these
steps are not necessary.

Orabug: 28782057

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>