oracle: Load RDMAIP kernel module using hot-plug
Load the 'resilient_rdmaip' kernel module using the upstream kernel-boot
hot-plug framework.

NB: In the past, changes were made to the network service
(/etc/rc.d/init.d/network) to make it aware of the RDMAIP module. Those
changes include:

1. Disabling active/active bonding when the network service is stopped
   and enabling active/active bonding when the network service is
   started. In this context, stopping the network service means running
   'ifdown' on all interfaces so it is not clear why RDMAIP would still
   be atempting to do anything with the interfaces when they are down.
   RDMAIP should use netlink or other facilities to be aware of the
   interface state.

2. Triggering RDMAIP to run its bonding code when interfaces are up.
   This is apparently done to avoid some sort of delay in how long it
   might otherwise take RDMAIP to bond the interfaces. Again, this
   should be done automatically by having RDMAIP use netlink to be aware
   of interfaces state changes.

At this time, these changes are not included.

For #1: The network service used previously is no longer being used.
Instead, network interface configuration is done using systemd-networkd.
The systemd-networkd service has no concept of starting and stopping the
network. It is a network interface configuration service and stopping it
does not change the state of any of the interfaces it manages. Under
systemd, it is not possible to stop a service or socket and have all of
the netork interfaces shutdown in the way that would happen in the past
with "service network stop". If it is necessary to start and stop RDMAIP
then this must be done based on the state of the interfaces that RDMAIP
is currently bonding and not on a systemd-wide basis.

For #2: Again, there is no single service that enables all interfaces or
disables all interfaces. Furthermore, the concept of the network being
"up" in the current version of systemd is that the network is "up" as
soon as a single non-loopback interface has a routable IP addresses and
network connectivity. The network online target does NOT indicate that
all interfaces are active. It is possible, in some situations, to use
a 'systemd-networkd-wait-online.service' to have a service unit wait on
a particular interface but this requires prior knowledge of which
interface(s) to wait on and hardcoding this information into a service
unit. RDMAIP should be made aware of the state of the interfaces it is
bonding and run its actions when both initially reach link up.

Orabug: 28782057

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>