ibacm: Fix improper refcnt
In the default provider, acmp, the port's sa_dest.refcnt is
initialized to one in acmp_init_dest(). When a port associated with
the device is added, the refcnt is set once again to one in
acmp_port_up().

For a system with VFs enabled, the ports for both the PF and the VFs
are added. This because ibacm uses the Node GUID, Port, and pkey
triple to designate an endpoint. Since the Node GUID is the very same
for the PF and VFs, we will get several port up events.

    acmp_port_up: mlx4_0 1
    acmp_port_up: mlx4_0 1 is up
    acmp_port_up: mlx4_0 1
    acmp_port_up: mlx4_0 1 is up

Now, when/if this ports goes down, acmp_port_down() will be called. It
decrements the refcnt and waits endlessly until it becomes zero. For
the first call to acmp_port_down(), it becomes zero, but the
subsequent call will loop forever in said busy-waiting loop, due to
the refcnt becoming minus one.

Fixed by changing the atomic_set() to atomic_inc() in acmp_port_up()
and test for one instead of the while loop in acmp_port_down().

Also added printing of instance number to ease debugging.

With the fix, the extract from the log file yields:

    acmp_port_up: mlx4_0 1
    acmp_port_up: mlx4_0 1 1 is up
    acmp_port_up: mlx4_0 1
    acmp_port_up: mlx4_0 1 2 is up
    acmp_port_down: mlx4_0 1
    acmp_port_down: mlx4_0 1 2 is down
    acmp_port_down: mlx4_0 1
    acmp_port_down: mlx4_0 1 1 is down

Signed-off-by: Håkon Bugge <haakon.bugge@oracle.com>
Reviewed-by: Mark Haywood <mark.haywood@oracle.com>

Orabug: 29037237

(cherry picked from commit 98523974c5bdfc695af36693af93f09e416a56eb)
cherry-pick-repo=linux-rdma/rdma-core.git
unmodified-from-upstream: 98523974c5bdfc695af36693af93f09e416a56eb

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Signed-off-by: Aron Silverton <aron.silverton@oracle.com>