ibacm: Allocate end-point addresses dynamically in provider
ibacm's default provider has a fixed number (4) of end-point
addresses. An endpoint is associated with the node GUID. Hence, if a
system is brought up with VFs enabled, the default provider will not
be able to handle the end-point addresses.

Also, if a port is assigned an IPv4 address and scope link-local and
global IPv6 addresses, and the device has two ports, six addresses
must be supported by the provider per end-point.

Now, consider the combination of multiple VFs instantiated and a
multitude of addresses assigned to each, it is hard to come up with a
hard limit.

Hence, we need to allocate the number of end-point addresses
dynamically in the provider.

acmp uses the address of a particular addr_info entry in struct
acmp_ep as its addr_context, stored in ibacm's data
structures. However, since these addresses now may change, due to
realloc(), we change the addr_context to a small struct containing a
pointer to the end-point and an addr_inx.

This way, we can safely get the correct address of a realloc'd
addr_info entry.

Signed-off-by: Håkon Bugge <haakon.bugge@oracle.com>

Orabug: 29582724

(cherry picked from commit 3a76a4963661aa03f3f581b94669e471dba2bec6)
cherry-pick-repo=github.com/linux-rdma/rdma-core.git
unmodified-from-upstream: 3a76a4963661aa03f3f581b94669e471dba2bec6

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>