mlx5: Introduce a wait queue for SRQ WQEs
When allocating the WQE buffer try to allocate more space than required.
The extra space will serve as a place for WQEs that were recently
switched from HW ownership to SW ownership to cool down before being
posted again. This is useful with WQEs with ODP buffers that were
consumed by HW but weren't handled yet by the page-fault handler in
kernel.

The policy of the wait queue is FIFO so a WQE gets out the wait queue
after N-1 WQEs get in when N is the size of the wait queue.

WQEs in the wait queue are considered to be in SW ownership except they
are not counted as candidates for posting. This means that WQEs in the
wait queue aren't in HW ownership while there.

Putting a WQE in the wait queue means that it's no longer available for
posting. When that happens, another WQE in the wait queue needs to be
taken out of there to replace it.

Having a wait queue is not mandatory. If the extra resources that are
required for the wait queue are beyond the limits of the system then the
SRQ will operate without a wait queue.

Signed-off-by: Moni Shoua <monis@mellanox.com>
Reviewed-by: Artemy Kovalyov <artemyko@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>

Orabug: 29318356

(cherry picked from commit 693cc37bbbbf89894c077a9ef20f1ba787ec238a)
cherry-pick-repo=linux-rdma/rdma-core.git

Conflicts:
	providers/mlx5/mlx5.h
	providers/mlx5/verbs.c

The conflict in mlx5.h is a simple merge issue between changes made
upstream and locally. We resolved them by accepting both the changes
upstream, adding some inline functions, and the local changes, adding
the prototypes for mlx5_get_legacy_xrc() and mlx5_set_legacy_xrc().

The conflict in verbs.c is because the upstream patch was
made against patch 2e559a9d already being committed upstream.
This patch does not exist locally yet. To resolve the
commit, we remove the change to the get_cqe_size()
signature.

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>