mlx5: Handle ODP fault completion in SRQ
A SRQ WQE that has ODP buffers might be completed with error and a
special syndrome. This indicates that the HW couldn't scatter the data
to the WQE buffers on one hand but needed to consume the WQE anyway.
This type of error shouldn't be raised to the poller of the CQ but
handled inside the driver.

The WQE for which the completions arrived for need to be re-posted to
keep the integrity of the SRQ from the application point of view. The
rep-posted WQE is taken from the SRQ head which means that the completed
WQE becomes free. To prevent it from being posted again with different
addresses and thus interfere the page-fault handler in the kernel, it is
recommended to put this WQE in the wait queue for cooling down.

Signed-off-by: Moni Shoua <monis@mellanox.com>
Reviewed-by: Artemy Kovalyov <artemyko@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>

Orabug: 29318356

(cherry picked from commit f3df671d32696be9d3e755986afbb470bc859f65)
cherry-pick-repo=github.com/linux-rdma/rdma-core.git
unmodified-from-upstream: f3df671d32696be9d3e755986afbb470bc859f65

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Acked-by: Aron Silverton <aron.silverton@oracle.com>