oracle: Add IPoIB interface configuration scripts
The RDMA Core project's initialization is intended to be systemd-only.
The scripts that provide the initialization are provided by the
rdma-core package. This package is analogous to the old rdma package but
drops most of the legacy SysV init infrastructure. The scripts that
configure IPoIB interfaces were inadvertently carried over into
rdma-core from rdma and eventually dropped in v17 of RDMA Core. Here is
the commit message:

------------------------------------------------------------------------
redhat/spec: Don't include ifup/ifdown-ib

When the original Red Hat spec file was created, many things that we
used to do in the old rdma package were brought over.  One of the things
brought over were the ifup-ib and ifdown-ib network scripts.  However,
those scripts are meaningless without the SysV-initscripts package which
includes all of the sysv network startup infrastructure.  The new
rdma-core was intended to be a systemd only service, and so we skipped
the Requirements: initscripts that we should have had in order to
support these scripts.  So, remove the scripts as they are old baggage.

Fixes: dac219f1da8b ("redhat/spec: build split rpm packages")
Signed-off-by: Doug Ledford <dledford@redhat.com>
------------------------------------------------------------------------

This assumed that there would be a systemd replacement, either natively
in systemd-networkd or systemd service unit files, to provide the
necessary interface configuration. This, however, is not yet the case.
This is similar to RoCE-specific network interface configuration in that
systemd-networkd does not natively support it. As such, we provided a
systemd service unit (interface-roce-config@.service) to provide that
functionality until it is natively supported. We had planned to do the
same for Infiniband-specific configuration that is not yet supported by
systemd-networkd.

However, Exadata has elected not to use systemd-networkd and we reverted
to SysV init methods for RoCE configuration. We need to do the same for
IPoIB by restoring the scripts:

    /etc/sysconfig/network-scripts/ifdown-ib
    /etc/sysconfig/network-scripts/ifup-ib

to the rdma-core package.

In fact, despite what the commit listed above from RedHat says, this is
what RedHat did in the package that they distribute. That is, their
distro-specific source directory in the RDMA Core repo removed the files
but the package that they distribute does include them. One can see this
in the rdma-core package OL distributes in ol7_latest as it is a rebuild
of the RHEL package.

The IPoIB interface configuration could be re-introduced to our package
by adding a systemd service unit like the one we created for RoCE but it
makes more sense, at this time, to stick with all SysV or all systemd
for network interface configuration. This will be revisited if/when
Exadata decides to move forward with systemd-networkd.

Redhat has restored these files as well with the following note:

------------------------------------------------------------------------
Upstream doesn't need them, but RHEL7.x does, as people do still bypass
NetworkManager and bring up interfaces using SysV-initscripts type
things.
------------------------------------------------------------------------

Orabug: 29307270

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>