libibverbs: shpd_pingpong.c ought to be WITHOUT_ORACLE_EXTENSIONS aware
When compiled with WITHOUT_ORACLE_EXTENSIONS defined, shpd_pingpong.c
fails to build. Add conditional code to the source file to print a
message that the progran is unsupported when built that way.

Also, as part of this build, we will fix the conditional include of
config.h which was broken since HAVE_CONFIG_H was never defined.

Orabug: 29437535

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Reviewed-by: Yuval Shaia <yuval.shaia@oracle.com>
Reviewed-by: Gerd Rausch <gerd.rausch@oracle.com>
---
v0 -> v1:
        * Fixed indentation
	* Fixed return syntax
v1 -> v2:
        * fprintf() errors to stderr