oracle: Add support for named PKEY devices
Add support for named pkey devices using the Oracle Linux
create_named_child interface. This decouples the IPoIB PKEY number and
interface name.

By default, interfaces with a PKEY are named such that the PKEY is
embedded in the name and the corresponding ifcfg-<int>.<PKEY> file. This
is the same convention used for VLAN interfaces.

The default behavior is undesirable for Exadata. Exadata needs to be
able to transparently change the PKEY after a system has been deployed.
Removing the PKEY from the interface name and configuration file allows
such changes.

This change was originally made in bug 19477360 for the old "rdma"
initialization scripts package. The change was made prior moving the
source for "rdma" from SVN to Git so there is no commit to cherry-pick.

The Oracle-specific create_named_child interface (bug 19064704) is
implemented in kernel commits:

    UEK4: f70a5ac6ed7eb212ec282fa71f2993c0e003628d
    UEK5: 79f6d772f0f3eb2335f3c56a04e231457b49b138

Orabug: 29636104

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>
Tested-by: Leo Agranonik <leo.agranonik@oracle.com>
Reviewed-by: Mukesh Kacker <mukesh.kacker@oracle.com>