ibacm: check provider file ends with .so extension
acm_open_providers() reads filenames and checks to see if the filenames
contain the .so (shared object) extension via strstr(). But this does
not verify that the extension is found at the end of the filename - only
that .so is a substring somewhere in the filename. This means filenames
of the sort "libibacmp.so.org" will be matched and loaded as providers.
The check should be modified to verify that the extension is at the end
of the filename.

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>

Orabug: 31048252

(cherry picked from commit ad5d934d688911149d795aee1d3b9fa06bf171a9)
cherry-pick-repo=github.com/linux-rdma/rdma-core.git
unmodified-from-upstream: ad5d934d688911149d795aee1d3b9fa06bf171a9

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>
Reviewed-by: Håkon Bugge <haakon.bugge@oracle.com>