srp_daemon.service: systemd Unknown lvalue messages in /var/log/messages
The srp_daemon systemd service configuration files contain hardening
options not yet supported by the systemd included in OL7. The options
MemoryDenyWriteExecute, ProtectControlGroups, ProtectKernelModules and
RestrictRealtime were added to systemd upstream in versions 231 and 232.

+--------------------------+---------------+
| lvalue                   | systemd       |
+--------------------------+---------------+
| MemoryDenyWriteExecute   | 231           |
| ProtectControlGroups     | 232           |
| ProtectKernelModules     | 232           |
| RestrictRealtime         | 231           |
+--------------------------+---------------+

The OL7 systemd is currently based off of upstream version 219, while
the systemd on OL8 is based off of upstream version 239. Given that
these options are not supported on OL7, we will remove them from the
configuration file as part of rpmbuild. This change can be reverted in
the future when/if the systemd on OL7 supports these options.

Orabug: 31377510

Signed-off-by: Mark Haywood <mark.haywood@oracle.com>